Moved photo preview swipe down gesture to screen level and fix scrub blink on loaded asset
